Remove terminal container view, switch to notify errors
